Can't use the parser in a web environment #64
* Removed usage of readFileSync
* Isolated the local file reader in a test class
* Replaced Node's path method for path-normalize (MIT)
* Bumped version from 1.0.2 to 2.0.0 because of API break
* Updated README.MD

As a side effect, also the node imports were organized.
