+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "@ibm/ibmi-eventf-parser",
   "description": "Parse event files that are generated by IBM i compilers",
-  "version": "1.0.2",
+  "version": "2.0.0",
   "publisher": "IBM",
   "author": "IBM",
   "license": "Apache-2.0",
@@ -26,6 +26,9 @@
     "test": "npx vitest run",
     "test:dev": "npx vitest",
     "coverage": "npx vitest run --coverage"
+  },
+  "dependencies": {
+    "path-normalize": "^6.0.13"
   },
   "devDependencies": {
     "@vitest/coverage-v8": "^1.0.2",
