+++ src/Parser.ts
@@ -2,7 +2,6 @@
  * (c) Copyright IBM Corp. 2023
  */
 
-import * as path from "path";
 import { IRecordT } from './record/IRecordT';
 import { TimestampRecord } from './record/TimestampRecord';
 import { ProcessorRecord } from './record/ProcessorRecord';
@@ -14,7 +13,7 @@ import { ErrorInformationRecord } from './record/ErrorInformationRecord';
 import { FileEndRecord } from './record/FileEndRecord';
 import { ExpansionRecord } from './record/ExpansionRecord';
 import { ExpansionProcessor } from './ExpansionProcessor';
-import { readFileSync } from 'fs';
+import normalize from 'path-normalize';
 
 class SourceFile {
   constructor(private location: string, private browseMode: boolean) { }
@@ -84,23 +83,19 @@ export class Parser {
   }
 
   private log(content: string) {
-    if (process.env.DEBUG) {
+    if (process && process.env.DEBUG) {
       console.log(content);
     }
   }
 
   /**
    * To parse an eventf generated by an IBM i compiler.  
-   * @param fileName local file containing the eventf encoded in UTF-8 
-   * @param markerCreator optionally passing in an object implementing IMarkerCreator
+   * @param fileReader a generic reader implementing {@link ISequentialFileReader}. 
+   * @param markerCreator optionally passing in an object implementing {@link IMarkerCreator}
    *   allows the parser to call this object back with each error.  The original source
    *   file name and error location and error message will be in the information provided
    *   on the callback
    */
-  parseFile(fileName: string, markerCreator?: IMarkerCreator) {
-    this.parse(new FileReader(fileName), markerCreator);
-  }
-
   parse(fileReader: ISequentialFileReader, markerCreator?: IMarkerCreator) {
     let word: string;
     let fileId: number;
@@ -381,7 +376,7 @@ export class Parser {
 
   private resolveRelativePath(location: string) {
     // Normalize the location in case the location is a relative path
-    location = path.normalize(location).toString();
+    location = normalize(location).toString();
 
     // After normalization, / will be change to \ for windows path,
     // so have to change it back.
@@ -428,27 +423,5 @@ export class Parser {
     } else {
       return [];
     }
-  }
-}
-
-export default class FileReader implements ISequentialFileReader {
-  file: string;
-  linesArray: string[];
-  currLineNum = 0;
-
-  constructor(fileName: string) {
-    this.file = readFileSync(fileName, 'utf-8');
-    this.linesArray = this.file.split(/\r?\n/)
-  }
-
-  public readNextLine() {
-    let line: string | undefined;
-    if (this.currLineNum < this.linesArray.length) {
-      line = this.linesArray[this.currLineNum];
-      this.currLineNum++;
-    } else {
-      line = undefined;
-    }
-    return line;
   }
 }
